Start to move enums and struct to modules
