+++ src/parse/nodes.rs
@@ -1,6 +1,7 @@
 mod blocs;
 mod main_nodes;
-mod expressions;
+pub(crate) mod expressions;
 mod if_else;
 mod fct;
 mod vars;
+mod classes;
