+++ src/parse/nodes/classes.rs
@@ -1,0 +1,6 @@
+use crate::parse::Scope;
+
+pub struct ClassDec {
+    identifier: String,
+    scope: Box<Scope>,
+}
